fix: Complete build fixes for RawrXD-Win32IDE
Build Fixes:
- Synchronized error_recovery_system.h with cpp implementation
- Synchronized hybrid_cloud_manager.h with cpp (added missing methods/members)
- Added missing slot implementations in hybrid_cloud_manager.cpp
- Fixed signal name systemHealthChanged -> systemHealthUpdated
- Added language field to GeneratedTest struct
- Fixed std::max type mismatch (int vs qsizetype)
- Fixed QString::replace on const reference in autonomous_model_manager
- Fixed most vexing parse with QNetworkRequest
- Added TestResult struct to benchmark_menu_widget.hpp
- Removed duplicate TestResult definition from cpp

CMakeLists.txt Updates:
- Added quant_utils to RawrXD-Win32IDE link libraries
- Added gguf_vocab_resolver.cpp to sources
- Added benchmark headers for Qt AutoMoc

Tokenizer Fixes:
- Enterprise BPE tokenizer with greedyLongestMatch fallback
- ReverseVocab lookup table for proper token encoding
- System prompt support for enterprise deployment

All targets now build successfully.

+++ src/performance_monitor.cpp
@@ -300,7 +300,7 @@ double PerformanceMonitor::getPercentile(const QString& component, const QString
     
     // Calculate percentile index
     int index = static_cast<int>((percentile / 100.0) * values.size());
-    index = std::min(index, values.size() - 1);
+    index = std::min(index, static_cast<int>(values.size() - 1));
     
     return values[index];
 }
